Changing data types to be `std::byte` in cuio

+++ cpp/src/io/orc/orc_field_writer.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2022, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -55,15 +55,15 @@ struct ProtobufWriter::ProtobufFieldWriter {
   {
     struct_size += p->put_uint(encode_field_number<std::vector<T>>(field));
     auto lpos = p->m_buf->size();
-    p->put_byte(0);
+    p->put_byte(static_cast<std::byte>(0));
     auto sz = std::accumulate(value.begin(), value.end(), 0, [p = this->p](size_t sum, auto val) {
       return sum + p->put_uint(val);
     });
 
     struct_size += sz + 1;
     for (; sz > 0x7f; sz >>= 7, struct_size++)
-      p->m_buf->insert(p->m_buf->begin() + (lpos++), static_cast<uint8_t>((sz & 0x7f) | 0x80));
-    (*(p->m_buf))[lpos] = static_cast<uint8_t>(sz);
+      p->m_buf->insert(p->m_buf->begin() + (lpos++), static_cast<std::byte>((sz & 0x7f) | 0x80));
+    (*(p->m_buf))[lpos] = static_cast<std::byte>(sz);
   }
 
   /**
@@ -85,12 +85,12 @@ struct ProtobufWriter::ProtobufFieldWriter {
   {
     struct_size += p->put_uint(encode_field_number(field, ProtofType::FIXEDLEN));
     auto lpos = p->m_buf->size();
-    p->put_byte(0);
+    p->put_byte(static_cast<std::byte>(0));
     auto sz = p->write(value);
     struct_size += sz + 1;
     for (; sz > 0x7f; sz >>= 7, struct_size++)
-      p->m_buf->insert(p->m_buf->begin() + (lpos++), static_cast<uint8_t>((sz & 0x7f) | 0x80));
-    (*(p->m_buf))[lpos] = static_cast<uint8_t>(sz);
+      p->m_buf->insert(p->m_buf->begin() + (lpos++), static_cast<std::byte>((sz & 0x7f) | 0x80));
+    (*(p->m_buf))[lpos] = static_cast<std::byte>(sz);
   }
 
   /**
